requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.2.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.31
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==3.19.2
